Moved AutoConstants to Constants package and moved auto config into CommandSwerveDrivetrain constructor.
